backend/pyproject.toml
[project]
name = "expense-tracker"
version = "0.1.0"
description = "AI-powered expense tracker with Telegram bot interface"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.42.0",
    "openai>=1.58.0",
    "python-telegram-bot[job-queue]>=21.0",
    "python-multipart>=0.0.18",
    "aiofiles>=24.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
